AcctIdx: move items to CachedBucket
