+++ runtime/src/bucket_map_holder.rs
@@ -1,11 +1,14 @@
 use crate::accounts_index::{AccountMapEntry, IndexValue};
 use crate::bucket_map_holder_stats::BucketMapHolderStats;
 use crate::waitable_condvar::WaitableCondvar;
+use solana_measure::measure::Measure;
 use solana_sdk::pubkey::Pubkey;
 use std::collections::HashMap;
-use std::sync::atomic::{AtomicBool, Ordering};
+use std::ops::RangeBounds;
+use std::sync::atomic::{AtomicBool, AtomicU64, Ordering};
 use std::sync::{Arc, RwLock};
 use std::time::Duration;
+type K = Pubkey;
 
 // will eventually hold the bucket map
 #[derive(Default)]
@@ -17,7 +20,7 @@ pub struct BucketMapHolder<T: IndexValue> {
 #[derive(Default)]
 pub struct CachedBucket<T: IndexValue> {
     _bin: usize,
-    _stats: Arc<BucketMapHolderStats>,
+    stats: Arc<BucketMapHolderStats>,
     pub map: RwLock<HashMap<Pubkey, AccountMapEntry<T>>>,
 }
 
@@ -25,9 +28,44 @@ impl<T: IndexValue> CachedBucket<T> {
     pub fn new(stats: Arc<BucketMapHolderStats>, bin: usize) -> Self {
         Self {
             _bin: bin,
-            _stats: stats,
+            stats,
             ..Self::default()
         }
+    }
+
+    pub fn items<R>(&self, range: &Option<&R>) -> Vec<(K, AccountMapEntry<T>)>
+    where
+        R: RangeBounds<Pubkey> + std::fmt::Debug,
+    {
+        Self::update_stat(&self.stats().items, 1);
+        let map = self.map().read().unwrap();
+        let mut result = Vec::with_capacity(map.len());
+        map.iter().for_each(|(k, v)| {
+            if range.map(|range| range.contains(k)).unwrap_or(true) {
+                result.push((*k, v.clone()));
+            }
+        });
+        result
+    }
+
+    fn map(&self) -> &RwLock<HashMap<Pubkey, AccountMapEntry<T>>> {
+        &self.map
+    }
+
+    fn stats(&self) -> &BucketMapHolderStats {
+        &self.stats
+    }
+
+    fn update_stat(stat: &AtomicU64, value: u64) {
+        if value != 0 {
+            stat.fetch_add(value, Ordering::Relaxed);
+        }
+    }
+
+    pub fn update_time_stat(stat: &AtomicU64, mut m: Measure) {
+        m.stop();
+        let value = m.as_us();
+        Self::update_stat(stat, value);
     }
 }
